+++ server/application/user/user-service.ts
@@ -2,7 +2,7 @@ import type { User, ProfileVisibility } from "@/server/domain/models/user/user";
 import type { UserId } from "@/server/domain/common/ids";
 import type { UserRepository } from "@/server/domain/models/user/user-repository";
 import type { createAccessService } from "@/server/application/authz/access-service";
-import type { RateLimiter } from "@/server/application/common/rate-limiter";
+import type { RateLimiter } from "@/server/domain/common/rate-limiter";
 import { BadRequestError, ForbiddenError } from "@/server/domain/common/errors";
 import { USER_PASSWORD_MAX_LENGTH } from "@/server/domain/models/user/user";
