+++ server/infrastructure/auth/nextauth-handler.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, test, vi } from "vitest";
 import type { JWT } from "next-auth/jwt";
-import type { RateLimiter } from "@/server/application/common/rate-limiter";
+import type { RateLimiter } from "@/server/domain/common/rate-limiter";
 import { TooManyRequestsError } from "@/server/domain/common/errors";
 import type { UserRepository } from "@/server/domain/models/user/user-repository";
 import { USER_NAME_MAX_LENGTH } from "@/server/domain/models/user/user";
