+++ server/infrastructure/rate-limit/noop-rate-limiter.ts
@@ -1,4 +1,4 @@
-import type { RateLimiter } from "@/server/application/common/rate-limiter";
+import type { RateLimiter } from "@/server/domain/common/rate-limiter";
 
 /**
  * 何もしない RateLimiter 実装。authorize コールバックが呼ばれないコンテキスト
